--- a/.travis.yml
+++ b/.travis.yml
@@ -1,7 +1,7 @@
 language: node_js
 cache: npm
 install:
-  - npm install
+  - npm ci
 before_script:
   - npm run start:test-server
 script: