--- a/Dockerfile
+++ b/Dockerfile
@@ -4,7 +4,8 @@ WORKDIR /app
 COPY ./package.json ./package-lock.json ./
 RUN npm install
 
-COPY . .
+COPY ./scripts ./scripts
+COPY ./widget ./widget
 RUN npm run build
 RUN npm prune --production
 
@@ -14,5 +15,6 @@ ENV DB_STORAGE /data/plusfries.db
 ENV LOG_LEVEL info
 ENV PORT 8080
 
+COPY . .
 HEALTHCHECK --interval=5m --timeout=3s CMD wget -qO- 0.0.0.0:${PORT}/healthcheck || exit 1
 CMD ["npm", "run", "start:prod"]