--- a/scripts/widget.sh
+++ b/scripts/widget.sh
@@ -1,17 +1,15 @@
-#!/bin/sh
+#!/bin/bash
 # Compress and minify widget
 set -e
 
-TEMP_DIR="$(mktemp -d)"
-mkdir -p ./public
+OUT_FILE="public/plusfries.sh"
+BIN="./node_modules/.bin"
 
-echo "start minifying..."
-# minify css
-npx postcss widget/plusfries.css >$TEMP_DIR/plusfries.css
-# compile and minify js app
-npx babel widget/plusfries.js --out-file $TEMP_DIR/plusfries.js
-npx terser $TEMP_DIR/plusfries.js -o $TEMP_DIR/plusfries.js -m -c
-# inject css
-CSS=$(cat $TEMP_DIR/plusfries.css)
-sed "s/PLUS_FRIES_CSS/${CSS}/g" $TEMP_DIR/plusfries.js >public/plusfries.js
+echo "minifying css..."
+CSS="$(time $BIN/uglifycss widget/plusfries.css)"
+echo "transpiling js..."
+ES5="$(time $BIN/babel widget/plusfries.js)"
+echo "minifying js..."
+COMPACT="$(time (echo "$ES5" | $BIN/terser -m -c))"
+echo "$COMPACT" | sed "s/PLUS_FRIES_CSS/${CSS}/g" >$OUT_FILE
 echo "minification done!"