Test:
```sh
seq 10 | while read; do time ./scripts/widget.sh >/dev/null; done
```

Before:
```sh
./scripts/widget.sh > /dev/null  2.28s user 0.67s system 77% cpu 3.812 total
./scripts/widget.sh > /dev/null  2.04s user 0.32s system 115% cpu 2.045 total
./scripts/widget.sh > /dev/null  2.01s user 0.31s system 114% cpu 2.025 total
./scripts/widget.sh > /dev/null  1.97s user 0.30s system 114% cpu 1.983 total
./scripts/widget.sh > /dev/null  1.96s user 0.31s system 114% cpu 1.977 total
./scripts/widget.sh > /dev/null  1.93s user 0.29s system 114% cpu 1.945 total
./scripts/widget.sh > /dev/null  2.00s user 0.31s system 114% cpu 2.007 total
./scripts/widget.sh > /dev/null  1.98s user 0.31s system 114% cpu 2.004 total
./scripts/widget.sh > /dev/null  2.03s user 0.32s system 114% cpu 2.053 total
./scripts/widget.sh > /dev/null  2.03s user 0.32s system 114% cpu 2.053 total
```

After:
```sh
./scripts/widget.sh > /dev/null  0.69s user 0.17s system 88% cpu 0.963 total
./scripts/widget.sh > /dev/null  0.55s user 0.10s system 107% cpu 0.605 total
./scripts/widget.sh > /dev/null  0.55s user 0.10s system 107% cpu 0.603 total
./scripts/widget.sh > /dev/null  0.55s user 0.10s system 107% cpu 0.605 total
./scripts/widget.sh > /dev/null  0.55s user 0.10s system 108% cpu 0.601 total
./scripts/widget.sh > /dev/null  0.55s user 0.10s system 108% cpu 0.604 total
./scripts/widget.sh > /dev/null  0.55s user 0.10s system 108% cpu 0.597 total
./scripts/widget.sh > /dev/null  0.55s user 0.10s system 108% cpu 0.599 total
./scripts/widget.sh > /dev/null  0.54s user 0.10s system 108% cpu 0.584 total
./scripts/widget.sh > /dev/null  0.54s user 0.10s system 107% cpu 0.595 total
```

Not writing files saved ~0.2s. Uglifycss saved the rest.